requirements.txt
flet
python-dotenv
mysql-connector-python
